added DynamicAvatar component to style the users profile picture
