+++ src/components/Appbar.jsx
@@ -7,53 +7,17 @@ import Typography from '@mui/material/Typography';
 import Menu from '@mui/material/Menu';
 import MenuIcon from '@mui/icons-material/Menu';
 import Container from '@mui/material/Container';
-import Avatar from '@mui/material/Avatar';
 import Button from '@mui/material/Button';
 import Tooltip from '@mui/material/Tooltip';
 import MenuItem from '@mui/material/MenuItem';
 import AdbIcon from '@mui/icons-material/Adb';
 import Divider from '@mui/material/Divider';
 import { Link as MuiLink, Stack } from '@mui/material';
+import DynamicAvatar from './DynamicAvatar';
 import { UserContext } from '../Context';
 
 const pages = ['Messages', 'Contacts'];
 const settings = ['Profile', 'Settings', 'Logout'];
-
-function stringToColor(string) {
-  let hash = 0;
-  let i;
-
-  /* eslint-disable no-bitwise */
-  for (i = 0; i < string.length; i += 1) {
-    hash = string.charCodeAt(i) + ((hash << 5) - hash);
-  }
-
-  let color = '#';
-
-  for (i = 0; i < 3; i += 1) {
-    const value = (hash >> (i * 8)) & 0xff;
-    color += `00${value.toString(16)}`.slice(-2);
-  }
-  /* eslint-enable no-bitwise */
-
-  return color;
-}
-
-function stringAvatar(name) {
-  let initials;
-  try {
-    initials = name.split(' ')[0][0] + name.split(' ')[1][0];
-  } catch (error) {
-    initials = name[0][0];
-  }
-
-  return {
-    sx: {
-      bgcolor: stringToColor(name),
-    },
-    children: `${initials}`,
-  };
-}
 
 function ResponsiveAppBar() {
   const user = React.useContext(UserContext);
@@ -189,7 +153,7 @@ function ResponsiveAppBar() {
             <Box sx={{ flexGrow: 0 }}>
               <Tooltip title="Open settings">
                 <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                  { user.username && <Avatar {...stringAvatar(user.username)} /> }
+                  { user.username && <DynamicAvatar name={user.username} /> }
                 </IconButton>
               </Tooltip>
               <Menu
@@ -221,7 +185,7 @@ function ResponsiveAppBar() {
                   p: 2,
                   alignItems: 'center',
                 }}>
-                  <Avatar {...stringAvatar(user.username)} />
+                  <DynamicAvatar name={user.username} />
                   <Typography 
                     variant='subtitle1'
                     component='p'
