+++ src/components/DynamicAvatar.jsx
@@ -1,0 +1,46 @@
+import React from 'react'
+import Avatar from '@mui/material/Avatar';
+
+function stringToColor(string) {
+  let hash = 0;
+  let i;
+
+  /* eslint-disable no-bitwise */
+  for (i = 0; i < string.length; i += 1) {
+    hash = string.charCodeAt(i) + ((hash << 5) - hash);
+  }
+
+  let color = '#';
+
+  for (i = 0; i < 3; i += 1) {
+    const value = (hash >> (i * 8)) & 0xff;
+    color += `00${value.toString(16)}`.slice(-2);
+  }
+  /* eslint-enable no-bitwise */
+
+  return color;
+}
+
+function stringAvatar(name) {
+  let initials;
+  try {
+    initials = name.split(' ')[0][0] + name.split(' ')[1][0];
+  } catch (error) {
+    initials = name[0][0];
+  }
+
+  return {
+    sx: {
+      bgcolor: stringToColor(name),
+    },
+    children: `${initials}`,
+  };
+}
+
+const DynamicAvatar = ({name}) => {
+  return (
+    <Avatar {...stringAvatar(name)} />
+  )
+}
+
+export default DynamicAvatar
